Remove balanced convolution sections.
