+++ src/LsNumerics/ConvolutionReverb.hpp
@@ -71,8 +71,6 @@ namespace LsNumerics
 
     namespace Implementation
     {
-        void SlotUsageTest();
-
         class AssemblyQueue
         {
             // single-reader, single-writer, designed to be friendly to the reader.
@@ -326,360 +324,17 @@ namespace LsNumerics
             std::vector<float> inputBuffer;
             std::vector<fft_complex_t> buffer;
         };
-        class CompiledButterflyOp
-        {
-        public:
-            CompiledButterflyOp(fft_index_t in0, fft_index_t in1, fft_index_t out, fft_index_t M_index)
-                : in0(in0), in1(in1), out(out), M_index(M_index)
-            {
-                assert(in0 != INVALID_INDEX);
-                assert(in1 != INVALID_INDEX);
-            }
-            CompiledButterflyOp(BinaryReader &);
-            void Tick(std::vector<fft_complex_t> &workingMemory)
-            {
-                fft_complex_t &M = workingMemory[M_index];
-                fft_complex_t t1 = workingMemory[in1] * M;
-                fft_complex_t t0 = workingMemory[in0];
-                workingMemory[out] = t0 + t1;
-                workingMemory[out + 1] = t0 - t1;
-            }
-            void Tick(fft_complex_t *workingMemory)
-            {
-                fft_complex_t &M = workingMemory[M_index];
-                fft_complex_t t1 = workingMemory[in1] * M;
-                fft_complex_t t0 = workingMemory[in0];
-                workingMemory[out] = t0 + t1;
-                workingMemory[out + 1] = t0 - t1;
-            }
-            void Write(BinaryWriter &writer) const;
-            fft_index_t in0, in1, out, M_index;
-#ifndef NDEBUG
-            std::string id;
-#endif
-        };
-        struct PlanStep
-        {
-        public:
-            using fft_float_t = double;
-            using fft_complex_t = std::complex<double>;
-
-            PlanStep() {}
-            PlanStep(BinaryReader &reader);
-
-            double Tick(double value, std::vector<fft_complex_t> &workingMemory)
-            {
-                workingMemory[inputIndex] = value;
-                size_t sz = ops.size();
-                for (std::size_t i = 0; i < sz; ++i)
-                {
-                    ops[i].Tick(workingMemory);
-                }
-                return workingMemory[this->outputIndex].real();
-            }
-
-            fft_complex_t Tick(fft_complex_t value, std::vector<fft_complex_t> &workingMemory)
-            {
-                fft_complex_t *p = &workingMemory[0];
-                p[inputIndex] = value;
-                for (std::size_t i = 0; i < ops.size(); ++i)
-                {
-                    ops[i].Tick(workingMemory);
-                }
-                return p[this->outputIndex];
-            }
-
-            void Write(BinaryWriter &writer) const;
-
-            fft_index_t inputIndex;
-            fft_index_t inputIndex2;
-            fft_index_t outputIndex;
-            std::vector<CompiledButterflyOp> ops;
-        };
-        class FftPlan
-        {
-        public:
-            static const char *MAGIC_FILE_STRING;
-            static constexpr uint64_t FILE_VERSION = 101;
-            static constexpr uint64_t MAGIC_TAIL_CONSTANT = 0x10394A2BE7F3C34D;
-
-            FftPlan(
-                std::size_t maxDelay,
-                std::size_t storageSize,
-                std::vector<PlanStep> &&ops,
-                std::size_t constantsOffset,
-                std::vector<fft_complex_t> &&constants,
-                size_t startingIndex,
-                size_t impulseFftOffset)
-                : norm(fft_float_t(1 / std::sqrt((double)ops.size()))),
-                  maxDelay(maxDelay),
-                  storageSize(storageSize),
-                  steps(std::move(ops)),
-                  constantsOffset(constantsOffset),
-                  constants(std::move(constants)),
-                  startingIndex(startingIndex),
-                  impulseFftOffset(impulseFftOffset)
-
-            {
-            }
-            FftPlan(BinaryReader &reader);
-
-            void Write(BinaryWriter &writer) const;
-
-            std::size_t Delay() const { return maxDelay; }
-            std::size_t Size() const { return steps.size(); }
-            std::size_t StorageSize() const { return storageSize; }
-            fft_float_t Norm() const { return norm; }
-            std::size_t StartingIndex() const { return startingIndex; }
-            std::size_t ImpulseFftOffset() const { return impulseFftOffset; }
-
-            double Tick(std::size_t step, double value, std::vector<fft_complex_t> &workingMemory)
-            {
-                return steps[step].Tick(value * norm, workingMemory);
-            }
-            fft_complex_t Tick(std::size_t step, fft_complex_t value, std::vector<fft_complex_t> &workingMemory)
-            {
-                return steps[step].Tick(value * norm, workingMemory);
-            }
-            float ConvolutionTick(std::size_t step, float value, std::vector<fft_complex_t> &workingMemory)
-            {
-                double t = (value * norm);
-                auto &planStep = steps[step];
-                workingMemory[planStep.inputIndex2] = t; // provide data for the extra inputs.
-                float result = (float)planStep.Tick(t, workingMemory);
-                return result;
-            }
-            void InitializeConstants(std::vector<fft_complex_t> &workingMemory)
-            {
-                for (size_t i = 0; i < constants.size(); ++i)
-                {
-                    workingMemory[i + constantsOffset] = constants[i];
-                }
-            }
-            void PrintPlan();
-            void PrintPlan(std::ostream &stream, bool trimIds);
-            void PrintPlan(const std::string &filename);
-            void CheckForOverwrites();
-
-            void ZeroOutput(size_t output, fft_index_t storageIndex)
-            {
-                size_t slot = (output + maxDelay) % steps.size();
-                steps[slot].outputIndex = storageIndex;
-            }
-
-        private:
-            double norm;
-            std::size_t maxDelay;
-            std::size_t storageSize;
-            std::vector<PlanStep> steps;
-            std::size_t constantsOffset;
-            std::vector<fft_complex_t> constants;
-            std::size_t startingIndex;
-            std::size_t impulseFftOffset;
-        };
-        using plan_ptr = std::shared_ptr<FftPlan>;
-
     }
-    /// @brief Serial DFT that requires the same computational expense for each sample.
-    ///
-    /// When performing an DFT of an audio stream using a block DFT implementation, the bulk of the comuptational expense occurs
-    /// every N samples where N is the size of the FFT. The balanced FFT incurs a fixed computational expense in each sample cycle,
-    /// while also making FFT results available earlier.
-
-    class BalancedFft
-    {
-    public:
-        using Plan = Implementation::FftPlan;
-        using plan_ptr = Implementation::plan_ptr;
-
-        BalancedFft(size_t size, FftDirection direction);
-        ~BalancedFft() {}
-
-        void PrintPlan();
-        void PrintPlan(const std::string &fileName) { plan->PrintPlan(fileName); }
-
-    public:
-        size_t Size() const { return plan->Size(); }
-        size_t Delay() const { return plan->Delay(); }
-
-        fft_complex_t Tick(fft_complex_t value)
-        {
-            fft_complex_t result = plan->Tick(planIndex, value, workingMemory);
-            if (++planIndex >= plan->Size())
-            {
-                planIndex = 0;
-            }
-            return result;
-        }
-        void Tick(std::size_t frames, fft_float_t *RESTRICT inputs, fft_complex_t *RESTRICT outputs)
-        {
-            for (size_t i = 0; i < frames; ++i)
-            {
-                outputs[i] = Tick(inputs[i]);
-            }
-        }
-        void Tick(std::size_t frames, fft_complex_t *RESTRICT inputs, fft_complex_t *RESTRICT outputs)
-        {
-            for (size_t i = 0; i < frames; ++i)
-            {
-                outputs[i] = Tick(inputs[i]);
-            }
-        }
-        void Tick(std::size_t frames, fft_complex_t *RESTRICT inputs, fft_float_t *RESTRICT outputs)
-        {
-            for (size_t i = 0; i < frames; ++i)
-            {
-                outputs[i] = Tick(inputs[i]).real();
-            }
-        }
-        void Reset();
-
-    private:
-        void SetPlan(plan_ptr plan);
-
-        std::vector<fft_float_t> inputBuffer;
-        std::vector<fft_complex_t> workingMemory;
-        plan_ptr plan;
-        size_t planIndex = 0;
-
-    private:
-        struct PlanKey
-        {
-            size_t size;
-            FftDirection direction;
-
-            bool operator==(const PlanKey &other) const
-            {
-                return size == other.size && direction == other.direction;
-            }
-            size_t hash() const
-            {
-                return (size << 1) ^ (size_t)direction;
-            }
-        };
-        struct PlanKeyHash
-        {
-            size_t operator()(const PlanKey &key) const
-            {
-                return key.hash();
-            }
-        };
-        static std::unordered_map<PlanKey, plan_ptr, PlanKeyHash> planCache;
-        static plan_ptr GetPlan(std::size_t size, FftDirection direction);
-    };
-
-    /// @brief Convolution section with balanced exection time per cycle.
-    /// For moderate sizes of N, generating the execution plan can take a significant amount of time.
-    ///
-    /// The recommended way to use this package is to pre-generate files containing the execution plan.
-    /// See @ref BalancedConvolution for details.
-
-    class BalancedConvolutionSection
-    {
-    public:
-        using Plan = Implementation::FftPlan;
-        using plan_ptr = Implementation::plan_ptr;
-        using Fft = StagedFft;
-
-        static void SetPlanFileDirectory(const std::filesystem::path &path)
-        {
-            planFileDirectory = path;
-        }
-
-        BalancedConvolutionSection(size_t size, size_t offset, const std::vector<float> &impulseResponse);
-
-        BalancedConvolutionSection(size_t size, const std::vector<float> &impulseResponse)
-            : BalancedConvolutionSection(size, 0, impulseResponse)
-        {
-        }
-        BalancedConvolutionSection(const std::filesystem::path &path,
-                                   size_t offset, const std::vector<float> &data);
-
-        void Save(const std::filesystem::path &path);
-
-        ~BalancedConvolutionSection() {}
-
-        static size_t GetSectionDelay(size_t size);
-
-        size_t Size() const { return size; }
-        size_t Delay() const { return plan->Delay() - plan->Size() / 2; }
-
-        void PrintPlan()
-        {
-            plan->PrintPlan();
-        }
-        void PrintPlan(const std::string &fileName)
-        {
-            plan->PrintPlan(fileName);
-        }
-        float Tick(float value)
-        {
-            auto result = plan->ConvolutionTick(planIndex, value, workingMemory);
-            if (++planIndex >= plan->Size())
-            {
-                planIndex = 0;
-            }
-
-            return (float)(result);
-        }
-        void Tick(std::size_t frames, float *RESTRICT inputs, float *RESTRICT outputs)
-        {
-            for (size_t i = 0; i < frames; ++i)
-            {
-                outputs[i] = Tick(inputs[i]);
-            }
-        }
-
-        void Reset();
-        static bool PlanFileExists(size_t size);
-        static void ClearPlanCache();
-
-    private:
-        static std::mutex planCacheMutex;
-
-        static std::filesystem::path GetPlanFilePath(size_t size);
-
-        void SetPlan(plan_ptr plan, size_t offset, const std::vector<float> &impulseData);
-
-        static std::filesystem::path planFileDirectory;
-
-        size_t size = 0;
-
-        std::vector<fft_complex_t> workingMemory;
-        plan_ptr plan;
-        size_t planIndex = 0;
-
-        static std::unordered_map<std::size_t, plan_ptr> planCache;
-        static plan_ptr GetPlan(std::size_t size);
-        static plan_ptr GetPlan(const std::filesystem::path &path);
-    };
 
     /// @brief Convolution using a roughly fixed execution time per cycle.
     ///
-    /// Normal convolution requires fft operations that can be enormously expensive every n
-    /// cycles. BalancedConvolution spreads out the execution time so that there is roughly
-    /// a fixed amount execution time per cycle.
-    ///
-    /// Generation execution plans can take a significant amount of time (tens or hundreds of seconds).
-    /// It is recommended that you configure BalancedConvolution to use pre-generated execution plans
-    /// generated at compile time, and stored in files.
-    ///
-    /// To use pregenerated plan files, follow these steps.
-    ///
-    /// 1. Generate the execution plans using the GenerateFftPlans executable a build time.
-    ///
-    ///       GenerateFftPlans <output-directory>
-    ///
-    /// 2. Copy these files into a fixed location at install time. (May be read-only).
-    ///
-    /// 3. Call BalancedConvolution::SetPlanFileDirector(<plan-file-directory>) to
-    ///    to cause execution plans to be loaded from disk instead of being generated at
-    ///    runtime.
-    ///
+    /// A convolution section is performed on the audio thread using non-FFT convolution just long enough
+    /// to allow FFT convolutions to be performed on background threads.
+
     class BalancedConvolution : private LocklessQueue::IDelayLineCallback
     {
     public:
-        /// @brief Convolution reverb with balanced execution cost per cycle.
+        /// @brief Convolution
         /// @param schedulerPolicy Scheduler policy. See remarks.
         /// @param size numer of samples of the impulse response to use.
         /// @param impulseResponse Impulse samples
@@ -690,12 +345,17 @@ namespace LsNumerics
         /// BalancedConvolution uses a set of service threads to calculate large convolution sections. The priority
         /// of the service threads must be below that of audio thread (but still very high). The scheduler policty
         /// determines how the thread priority is set. If Scheduler::Realtime, the worker threads' scheduler policy
-        /// is set to SCHED_RR (realtime), and the prioriy of the threads are set to basePrioriy+1, basePriority+2,
-        /// &c. If Scheduler::Normal, the thread priorities are set using nice (3).
-        ///
-        /// Note that when running in realtime, the schedulerPolicy should always be SchedulerPolicy::Realtime.
+        /// is set to SCHED_RR (realtime). Actually priorities are chosen to provide optimal priorities for Linux
+        /// audio systems. Very large FFTs are schedule below +6 inorder not to interfere with USB audio services
+        /// which run at RT priority +6.
+        /// If Scheduler::Normal is specified, the thread priorities are set using nice (3). When running in 
+        /// realtime, the schedulerPolicy should always be SchedulerPolicy::Realtime.
         /// SchedulerPolicy::Normal allows unit testing in an environment where the running process may not have
-        /// suffieienctprivileges to set a realtime thread priority.
+        /// sufficient privileges to set a realtime thread priority, or perhaps processing offline.
+        ///
+        /// An dedicated thread (the assembly thread) is responsibile for assembling convolution sections calculated on
+        /// the background thread into a single stream for consumption by the audio thread. This ensures that (hopefully rare) 
+        /// system calls required to wait for background data don't execute on the audio thread. 
         ///
         /// The sampleRate and maxAudioBufferSize parameters are used to control scheduling of the background convolution
         /// sections. Note that sampleRate is the sample rate at which the audio thread is running, NOT the sample
@@ -703,13 +363,19 @@ namespace LsNumerics
         /// set to a modest value (e.g. 16..256), since the cost of running balanced convolution sections on the
         /// audio thread is significantly higher than running convolution sections in the background, but the
         /// realtime audio thread mut build up a lead-time significant enough to survive scheduling jitter on
-        /// the background audio threads. A high value of maxAudioBufferSize increases the lead-time required.
+        /// the background audio threads. A high value of maxAudioBufferSize increases the lead-time required,
+        /// and consequently, the amount of non-FFT processing that has to be done on the real-time thread.
+        ///
+        /// If determined to run with large buffer sizes, it would be a good idea to modifiy the realtime processing
+        /// thread to use short FFT-sections when processing in realtime. This is not currently implemented. The 
+        /// current implementation runs reasonable efficiently with buffer sizes less that 256 frames, and may well
+        /// behave badly with buffer sizes of 1024.
         ///
         BalancedConvolution(
             SchedulerPolicy schedulerPolicy,
             size_t size, const std::vector<float> &impulseResponse,
-            size_t sampleRate = 44100,
-            size_t maxAudioBufferSize = 128);
+            size_t sampleRate,
+            size_t maxAudioBufferSize);
 
         BalancedConvolution(
             SchedulerPolicy schedulerPolicy,
@@ -721,10 +387,7 @@ namespace LsNumerics
         }
 
         ~BalancedConvolution();
-        static void SetPlanFileDirectory(const std::filesystem::path &path)
-        {
-            BalancedConvolutionSection::SetPlanFileDirectory(path);
-        }
+
         size_t GetUnderrunCount() const { return (size_t)underrunCount; }
 
     private:
@@ -742,22 +405,8 @@ namespace LsNumerics
         float TickUnsynchronized(float value, float backgroundValue)
         {
             audioThreadToBackgroundQueue.Write(value);
-            double result = backgroundValue;
-            result += audioThreadToBackgroundQueue.DirectConvolve(directImpulse);
-            // for (size_t i = 0; i < directConvolutionLength; ++i)
-            // {
-            //     result += audioThreadToBackgroundQueue[i] * (double)directImpulse[i];
-            // }
-            for (auto &section : balancedSections)
-            {
-                result += section.fftSection.Tick(audioThreadToBackgroundQueue[section.sampleDelay]);
-            }
-            // moved to assembly thread.
-            // for (auto &sectionThread : directSectionThreads)
-            // {
-            //     result += sectionThread->Tick();
-            // }
-            return (float)result;
+
+            return (float)(backgroundValue + audioThreadToBackgroundQueue.DirectConvolve(directImpulse));
         }
 
     public:
@@ -934,21 +583,14 @@ namespace LsNumerics
         AudioThreadToBackgroundQueue audioThreadToBackgroundQueue;
         size_t directConvolutionLength;
 
-        struct Section
-        {
-            size_t sampleDelay;
-            BalancedConvolutionSection fftSection;
-        };
-
-        std::vector<Section> balancedSections;
         std::vector<DirectSection> directSections;
     };
 
     class ConvolutionReverb
     {
     public:
-        ConvolutionReverb(SchedulerPolicy schedulerPolicy, size_t size, const std::vector<float> &impulse)
-            : convolution(schedulerPolicy, size == 0 ? 0 : size - 1, impulse) // the last value is recirculated.
+        ConvolutionReverb(SchedulerPolicy schedulerPolicy, size_t size, const std::vector<float> &impulse,size_t sampleRate, size_t maxBufferSize)
+            : convolution(schedulerPolicy, size == 0 ? 0 : size - 1, impulse, sampleRate, maxBufferSize) // the last value is recirculated.
         {
             directMixDezipper.To(0, 0);
             reverbMixDezipper.To(1.0, 0);
